Tutorial1/requirements.txt
langchain-groq
python-dotenv
langchain-core
langchain-community
streamlit
